tests, cmake broken

+++ src/ecs/components/point_light.hpp
@@ -1,8 +1,8 @@
-#include "core/math/color.hpp"
+#include "math/color.hpp"
 
 // requires Spatial, position defined by the spatial
 struct PointLight {
-  double intensity; // w/m^2
-  double range; // m
+  float intensity; // w/m^2
+  float range; // m
   Color color;
 };
